mtg_manager/requirements.txt
requests>=2.28.0
Pillow>=9.0.0
pytesseract>=0.3.10
opencv-python>=4.6.0
easyocr>=1.6.0